Fix migration settings bundle budget
