+++ docs/spec-coverage.md
@@ -41,7 +41,7 @@ diagrams line up so future changes can see what must be updated.
 | Feature or subsystem | Source evidence | Canonical docs | Test evidence | Diagram coverage |
 |---|---|---|---|---|
 | Server boot, config, sessions, security headers | `server.ts`, `src/config/index.ts`, `src/middleware/auth.ts`, `src/middleware/csrf.ts`, `src/middleware/security.ts`, `src/types/session-file-store.d.ts`, `src/utils/logger.ts` | `docs/spec-server-security.md`, `docs/spec-api-endpoints.md`, `docs/spec-deployment.md`, `README.md`, `ONBOARDING.md` | `test/auth.test.ts`, `test/graceful-shutdown.test.ts`, `test/frontendRoutes.test.ts`, `test/logger.test.ts` | Auth/session flow below; startup sequence in `docs/spec-server-security.md`. |
-| Data-root export/import migration | `src/contracts/dataMigration.ts`, `src/services/dataMigrationService.ts`, `src/routes/chat/dataMigrationRoutes.ts`, `src/routes/chat.ts`, `server.ts`, `web/AgentCockpitWeb/src/api.js`, `web/AgentCockpitWeb/src/screens/settingsScreen.jsx` | `docs/spec-api-endpoints.md`, `docs/spec-data-models.md`, `docs/spec-backend-services.md`, `docs/spec-server-security.md`, `docs/spec-frontend.md`, `docs/spec-mobile-pwa.md`, `docs/reference/data-layout.md`, `docs/adr/0074-use-data-root-bundles-for-migration.md` | `test/dataMigrationService.test.ts`, `test/chat.dataMigration.test.ts`, `test/frontendRoutes.test.ts` | Startup sequence in `docs/spec-server-security.md`; frontend/client flow below. |
+| Data-root export/import migration | `src/contracts/dataMigration.ts`, `src/services/dataMigrationService.ts`, `src/routes/chat/dataMigrationRoutes.ts`, `src/routes/chat.ts`, `server.ts`, `web/AgentCockpitWeb/src/api.js`, `web/AgentCockpitWeb/src/screens/settingsScreen.jsx`, `web/AgentCockpitWeb/src/screens/settingsMigrationTab.jsx` | `docs/spec-api-endpoints.md`, `docs/spec-data-models.md`, `docs/spec-backend-services.md`, `docs/spec-server-security.md`, `docs/spec-frontend.md`, `docs/spec-mobile-pwa.md`, `docs/reference/data-layout.md`, `docs/adr/0074-use-data-root-bundles-for-migration.md` | `test/dataMigrationService.test.ts`, `test/chat.dataMigration.test.ts`, `test/frontendRoutes.test.ts` | Startup sequence in `docs/spec-server-security.md`; frontend/client flow below. |
 | Conversation, workspace, session, message, queue, archive, unread, usage ledger, usage pricing | `src/services/chatService.ts`, `src/services/chat/messageQueueStore.ts`, `src/services/chat/usageLedgerStore.ts`, `src/services/usagePricing/*`, `src/services/chat/artifactStore.ts`, `src/services/chat/workspaceIdentityStore.ts`, `src/services/chat/workspaceFeatureSettingsStore.ts`, `src/types/index.ts`, `src/contracts/*.ts`, `src/contracts/workspaces.ts`, `src/routes/chat.ts`, `src/routes/chat/conversationRoutes.ts`, `src/routes/chat/workspaceLocationRoutes.ts`, `src/routes/chat/statusRoutes.ts`, `src/utils/atomicWrite.ts`, `src/utils/keyedMutex.ts` | `docs/spec-data-models.md`, `docs/spec-api-endpoints.md`, `docs/spec-backend-services.md`, `docs/spec-deployment.md`, `docs/adr/0069-estimate-usage-costs-from-persisted-pricing-catalog.md`, `docs/adr/0072-use-per-conversation-git-worktrees-for-workspace-isolation.md`, `docs/adr/0073-use-stable-workspace-identities.md` | `test/chatContracts.test.ts`, `test/chatService.*.test.ts`, `test/chat.*.test.ts`, `test/chat.worktreeIsolation.test.ts`, `test/chat.usageLedgerStore.test.ts`, `test/usagePricing.test.ts`, `test/usagePricingStore.test.ts`, `test/chat.artifactStore.test.ts`, `test/chat.workspaceFeatureSettingsStore.test.ts`, `test/chat.workspaceIdentityStore.test.ts`, `test/messageQueue.test.ts`, `test/sessionStore.test.ts`, `test/usageProjection.test.ts`, `test/utils.*.test.ts` | Chat lifecycle below. |
 | Streaming, WebSocket replay, abort, durable stream jobs, finalizers | `src/ws.ts`, `src/services/streamJobRegistry.ts`, `src/services/streamJobSupervisor.ts`, `src/services/sessionFinalizerQueue.ts`, `src/routes/chat.ts`, `src/routes/chat/streamRoutes.ts`, `src/routes/chat/goalRoutes.ts` | `docs/spec-api-endpoints.md`, `docs/spec-backend-services.md`, `docs/spec-data-models.md` | `test/chat.websocket.test.ts`, `test/chat.streaming.test.ts`, `test/chat.rest.test.ts`, `test/sessionFinalizerQueue.test.ts`, `test/streamStore.test.ts` | Chat lifecycle below. |
 | Backend adapters and CLI profiles | `src/services/backends/base.ts`, `src/services/backends/registry.ts`, `src/services/backends/toolUtils.ts`, `src/services/backends/claudeCode.ts`, `src/services/backends/kiro.ts`, `src/services/backends/codex.ts`, `src/services/cliProfiles.ts`, `src/services/cliProfileAuthService.ts` | `docs/spec-backend-services.md`, `docs/spec-api-endpoints.md`, `docs/spec-data-models.md`, `BACKENDS.md`, `CLI_PROFILES_MULTI_ACCOUNT_PLAN.md` | `test/backends.test.ts`, `test/toolUtils.test.ts`, `test/kiroBackend.test.ts`, `test/codexBackend.test.ts`, `test/chat.cliProfileAuth.test.ts` | Adapter dispatch flow below. |
