+++ scripts/check-web-bundle-size.js
@@ -13,12 +13,13 @@ const DEFAULT_BUDGETS = {
   maxJsAsset: 230 * KIB,
   maxCssAsset: 230 * KIB,
   assets: {
-    'index.js': 210 * KIB,
+    'index.js': 211 * KIB,
     'react-vendor.js': 230 * KIB,
     'markdown-vendor.js': 160 * KIB,
     'kbBrowser.js': 115 * KIB,
     'workspaceSettings.js': 95 * KIB,
     'settingsScreen.js': 75 * KIB,
+    'settingsMigrationTab.js': 20 * KIB,
     'dialog.js': 60 * KIB,
     'filesBrowser.js': 45 * KIB,
     'memoryReview.js': 35 * KIB,
