+++ test/frontendRoutes.test.ts
@@ -611,6 +611,7 @@ describe('frontend routes', () => {
   test('Settings exposes data migration export/import controls', () => {
     const apiSrc = fs.readFileSync(path.join(ROOT, 'web/AgentCockpitWeb/src/api.js'), 'utf8');
     const settingsSrc = fs.readFileSync(path.join(ROOT, 'web/AgentCockpitWeb/src/screens/settingsScreen.jsx'), 'utf8');
+    const migrationTabSrc = fs.readFileSync(path.join(ROOT, 'web/AgentCockpitWeb/src/screens/settingsMigrationTab.jsx'), 'utf8');
     const cssSrc = fs.readFileSync(path.join(ROOT, 'web/AgentCockpitWeb/src/app.css'), 'utf8');
 
     expect(apiSrc).toContain("chatFetch('migration/status')");
@@ -630,25 +631,26 @@ describe('frontend routes', () => {
     expect(apiSrc).toContain("chatFetch(\n      'migration/import/confirm'");
     expect(apiSrc).toContain("chatFetch(`migration/checks${deep ? '?deep=true' : ''}`)");
     expect(settingsSrc).toContain("{ id: 'migration', label: 'Migration' }");
-    expect(settingsSrc).toContain('function MigrationTab()');
-    expect(settingsSrc).toContain("confirmation !== 'REPLACE'");
-    expect(settingsSrc).toContain('Replace all Agent Cockpit data?');
-    expect(settingsSrc).toContain('Import replaces everything in this installation.');
-    expect(settingsSrc).toContain('AgentApi.settings.startMigrationExport()');
-    expect(settingsSrc).toContain('AgentApi.settings.migrationExportJob(job.jobId)');
-    expect(settingsSrc).toContain('AgentApi.settings.migrationExportJobDownloadUrl(job.jobId)');
-    expect(settingsSrc).toContain('migration-button-progress');
-    expect(settingsSrc).toContain('hasMeasuredImportProgress');
-    expect(settingsSrc).toContain('Uploading…');
-    expect(settingsSrc).toContain('function MigrationImportProgress');
-    expect(settingsSrc).toContain('Uploading ${percent}%');
-    expect(settingsSrc).toContain('Processing');
-    expect(settingsSrc).toContain('Restoring');
-    expect(settingsSrc).toContain('function MigrationChecks');
-    expect(settingsSrc).toContain('Run checks help');
-    expect(settingsSrc).toContain('Deep checks help');
-    expect(settingsSrc).toContain('They do not contact Ollama.');
-    expect(settingsSrc).toContain('also contact each configured Ollama embedding host/model');
+    expect(settingsSrc).toContain("React.lazy(() => import('./settingsMigrationTab.jsx')");
+    expect(migrationTabSrc).toContain('export function MigrationTab()');
+    expect(migrationTabSrc).toContain("confirmation !== 'REPLACE'");
+    expect(migrationTabSrc).toContain('Replace all Agent Cockpit data?');
+    expect(migrationTabSrc).toContain('Import replaces everything in this installation.');
+    expect(migrationTabSrc).toContain('AgentApi.settings.startMigrationExport()');
+    expect(migrationTabSrc).toContain('AgentApi.settings.migrationExportJob(job.jobId)');
+    expect(migrationTabSrc).toContain('AgentApi.settings.migrationExportJobDownloadUrl(job.jobId)');
+    expect(migrationTabSrc).toContain('migration-button-progress');
+    expect(migrationTabSrc).toContain('hasMeasuredImportProgress');
+    expect(migrationTabSrc).toContain('Uploading…');
+    expect(migrationTabSrc).toContain('function MigrationImportProgress');
+    expect(migrationTabSrc).toContain('Uploading ${percent}%');
+    expect(migrationTabSrc).toContain('Processing');
+    expect(migrationTabSrc).toContain('Restoring');
+    expect(migrationTabSrc).toContain('function MigrationChecks');
+    expect(migrationTabSrc).toContain('Run checks help');
+    expect(migrationTabSrc).toContain('Deep checks help');
+    expect(migrationTabSrc).toContain('They do not contact Ollama.');
+    expect(migrationTabSrc).toContain('also contact each configured Ollama embedding host/model');
     expect(cssSrc).toContain('.migration-panel');
     expect(cssSrc).toContain('.migration-progress-button');
     expect(cssSrc).toContain('.migration-button-progress');
